niffler-e-2-e-tests-python/requirements.txt
pytest
requests
selene
faker
python-dotenv